test(e2e): dedupe oracle tool-call mock server
